pyproject.toml
[project]
name = "files-to-prompt"
version = "0.4"
description = "Concatenate a directory full of files into a single prompt for use with LLMs"
readme = "README.md"
authors = [{name = "Simon Willison"}]
license = {text = "Apache-2.0"}
requires-python = ">=3.9"
classifiers = [
    "License :: OSI Approved :: Apache Software License"
]
dependencies = [
    "grep-ast>=0.4.1",
    "tree-sitter>=0.21.3",
    "tree-sitter-languages>=1.10.2",
    "typer-slim>=0.15.1",
]

[project.urls]
Homepage = "https://github.com/simonw/files-to-prompt"
Changelog = "https://github.com/simonw/files-to-prompt/releases"
Issues = "https://github.com/simonw/files-to-prompt/issues"
CI = "https://github.com/simonw/files-to-prompt/actions"

[project.entry-points.console_scripts]
files-to-prompt = "files_to_prompt.cli:main"

[project.optional-dependencies]
test = ["pytest"]
